translator/pyproject.toml
[tool.poetry]
name = "text-translator"
version = "0.1.0"
description = "A screen OCR translator powered by Large Language Models (LLMs)"
authors = ["Qlefia <your.email@example.com>"]
readme = "README.md"
packages = [{include = "translator"}]

[tool.poetry.dependencies]
python = "^3.11"
pytesseract = "^0.3.10"
Pillow = "^10.1.0"
requests = "^2.31.0"
PyQt5 = "^5.15.9"
opencv-python = "^4.8.1.78"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 127
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 127

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true 